Refactor app.py: extract presentation, auth, retry logic, and simplify initialization
